Passing the user data along to every ejs file we render
Passing the user data to any rendered ejs file is necessary so that the navbar knows wether to show the login button or the logout button. Another plus for passing along user data along to ejs files is it opens the door to the possibility of custom tayloring welcome messages for the user.