Fixed date going back in time by one day.
For some reason, creating a date with this string yielded a date that was behind by one day:
"2020-12-02"

But, this one worked perfectly:
"2020-12-2"

I had to use some string magic to seperate the day and use the parseInt() method to get rid of the zero. Then, put it all back together, and it worked!